Refactoring (signature functions moved to one place) + constructors inheritance
